press-01-pencil
Blockchain tech firm, Pencil Works, is taking a stand against plagiarism and theft of intellectual property within the Handshake community.
Over the past few months, the company has been engaging with the Spanish community and posing questions about the significance of software licenses. Unfortunately, it was discovered that some members of the community wants to take advantage of the open-source nature of the code in the future, using it for commercial purposes without giving credit to the developers.
This issue came to a head when a member of the community was found to be using logos that were similar to those of established companies, including trademarks that were not their own. Pencil Works recognizes the importance of protecting the integrity of the community and its members, and therefore will not be releasing its code publicly. However, the company is committed to educating the public and showcasing their work in other ways on the internet.
The company is aware of the importance of community growth but believes that it is essential to protect its members and themselves. By preserving their code for their secure clients and company's future, Pencil Works is taking a responsible approach to open-source technology and setting a positive example for the Handshake community and beyond.